requirements.txt
# Medical ICD Mapper - Unified Requirements
# Install: pip install -r requirements.txt

# --- Backend: Core Framework ---
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# --- Backend: Database ---
sqlalchemy==2.0.25
alembic==1.13.1

# --- Backend: Validation & Config ---
pydantic>=2.12.5
pydantic-settings>=2.1.0
python-dotenv==1.0.0

# --- Backend: PDF Processing ---
pdfplumber==0.10.3
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0

# --- Backend: Medical NLP ---
spacy==3.6.1
scispacy==0.5.3
negspacy==1.0.4

# --- Backend: LLM (Gemini) ---
google-genai>=1.68.0
openai>=1.30.0

# --- Backend: Utilities ---
tenacity==8.2.3
tqdm==4.66.1
python-json-logger==2.0.7
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# --- Frontend: Streamlit Dashboard ---
streamlit==1.31.0
requests==2.31.0
pandas==2.2.0
plotly==5.18.0

# --- Dev / Testing ---
pytest==8.0.0
pytest-asyncio==0.23.3
httpx>=0.28.1
black==24.1.1
flake8==7.0.0

# --- Production Deployment ---
gunicorn==21.2.0
psycopg2-binary==2.9.9
en_core_sci_md @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.3/en_core_sci_md-0.5.3.tar.gz
